+++ android/engine/src/main/java/org/smartregister/fhircore/engine/ui/questionnaire/QuestionnaireActivity.kt
@@ -29,11 +29,15 @@ import androidx.lifecycle.lifecycleScope
 import ca.uhn.fhir.context.FhirContext
 import com.google.android.fhir.datacapture.QuestionnaireFragment
 import com.google.android.fhir.datacapture.QuestionnaireFragment.Companion.BUNDLE_KEY_QUESTIONNAIRE
+import com.google.android.fhir.datacapture.validation.QuestionnaireResponseValidator
 import com.google.android.fhir.logicalId
 import org.hl7.fhir.r4.model.Patient
 import org.hl7.fhir.r4.model.Questionnaire
 import org.hl7.fhir.r4.model.QuestionnaireResponse
 import org.smartregister.fhircore.engine.R
+import org.smartregister.fhircore.engine.ui.base.AlertDialogue.showConfirmAlert
+import org.smartregister.fhircore.engine.ui.base.AlertDialogue.showErrorAlert
+import org.smartregister.fhircore.engine.ui.base.AlertDialogue.showProgressAlert
 import org.smartregister.fhircore.engine.ui.base.BaseMultiLanguageActivity
 import org.smartregister.fhircore.engine.util.DefaultDispatcherProvider
 import org.smartregister.fhircore.engine.util.DispatcherProvider
@@ -71,6 +75,8 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
       showToast(getString(R.string.error_loading_form))
       finish()
     }
+
+    val loadProgress = showProgressAlert(this, R.string.loading)
 
     clientIdentifier = intent.getStringExtra(QUESTIONNAIRE_ARG_PATIENT_KEY)
 
@@ -111,7 +117,8 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
                 clientIdentifier == null ->
                   bundleOf(Pair(BUNDLE_KEY_QUESTIONNAIRE, parsedQuestionnaire))
                 clientIdentifier != null -> {
-                  //                  TODO it is not working. Takes forever to load form first time
+                  //                  TODO it is not working. Takes forever to load form first
+                  // time
                   //                  val parsedQuestionnaireResponse =
                   //                    parser.encodeResourceToString(
                   //
@@ -127,6 +134,7 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
           }
         supportFragmentManager.commit { add(R.id.container, fragment, QUESTIONNAIRE_FRAGMENT_TAG) }
       }
+      loadProgress.dismiss()
     }
 
     findViewById<Button>(R.id.btn_save_client_info).setOnClickListener(this)
@@ -140,51 +148,80 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
 
   override fun onClick(view: View) {
     if (view.id == R.id.btn_save_client_info) {
-      val questionnaireFragment =
-        supportFragmentManager.findFragmentByTag(QUESTIONNAIRE_FRAGMENT_TAG) as
-          QuestionnaireFragment
-      handleQuestionnaireResponse(questionnaireFragment.getQuestionnaireResponse())
+      showConfirmAlert(
+        context = this,
+        message = R.string.questionnaire_alert_submit_message,
+        title = R.string.questionnaire_alert_submit_title,
+        confirmButtonListener = { handleQuestionnaireSubmit() },
+        confirmButtonText = R.string.questionnaire_alert_submit_button_title
+      )
     } else {
       showToast(getString(R.string.error_saving_form))
     }
   }
 
-  open fun handleQuestionnaireResponse(questionnaireResponse: QuestionnaireResponse) {
-    if (questionnaire != null) {
-      val alertDialog = showDialog()
-
-      questionnaireViewModel.extractionProgress.observe(
-        this,
-        { result ->
-
-          // TODO: Unregister this observer
-
-          if (result) {
-            alertDialog.dismiss()
-            finish()
-          } else {
-            Timber.e("An error occurred during extraction")
-          }
-        }
-      )
-
-      questionnaireViewModel.extractAndSaveResources(
-        context = this@QuestionnaireActivity,
-        questionnaire = questionnaire!!,
-        questionnaireResponse = questionnaireResponse,
-        resourceId = intent.getStringExtra(QUESTIONNAIRE_ARG_PATIENT_KEY)
-      )
-    }
+  fun getQuestionnaireResponse(): QuestionnaireResponse {
+    val questionnaireFragment =
+      supportFragmentManager.findFragmentByTag(QUESTIONNAIRE_FRAGMENT_TAG) as QuestionnaireFragment
+    return questionnaireFragment.getQuestionnaireResponse()
   }
 
-  fun showDialog(): AlertDialog {
-    val dialogBuilder =
-      AlertDialog.Builder(this).apply {
-        setView(R.layout.dialog_saving)
-        setCancelable(false)
-      }
+  private lateinit var saveProcessingAlertDialog: AlertDialog
 
-    return dialogBuilder.create().apply { show() }
+  fun dismissSaveProcessing() {
+    saveProcessingAlertDialog.dismiss()
+  }
+
+  open fun handleQuestionnaireSubmit() {
+    saveProcessingAlertDialog = showProgressAlert(this, R.string.saving_registration)
+
+    val questionnaireResponse = getQuestionnaireResponse()
+
+    if (!validQuestionnaireResponse(questionnaireResponse)) {
+      saveProcessingAlertDialog.dismiss()
+
+      showErrorAlert(
+        this,
+        R.string.questionnaire_alert_invalid_message,
+        R.string.questionnaire_alert_invalid_title
+      )
+      return
+    }
+
+    handleQuestionnaireResponse(questionnaireResponse)
+
+    questionnaireViewModel.extractionProgress.observe(
+      this,
+      { result ->
+        if (result) {
+          finish()
+        } else {
+          Timber.e("An error occurred during extraction")
+        }
+
+        saveProcessingAlertDialog.dismiss()
+      }
+    )
+  }
+
+  fun validQuestionnaireResponse(questionnaireResponse: QuestionnaireResponse): Boolean {
+    return QuestionnaireResponseValidator.validate(
+        questionnaire.item,
+        questionnaireResponse.item,
+        this
+      )
+      .values
+      .flatten()
+      .all { it.isValid }
+  }
+
+  open fun handleQuestionnaireResponse(questionnaireResponse: QuestionnaireResponse) {
+    questionnaireViewModel.extractAndSaveResources(
+      context = this@QuestionnaireActivity,
+      questionnaire = questionnaire!!,
+      questionnaireResponse = questionnaireResponse,
+      resourceId = intent.getStringExtra(QUESTIONNAIRE_ARG_PATIENT_KEY)
+    )
   }
 
   companion object {
@@ -213,6 +250,12 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
   }
 
   override fun onBackPressed() {
-    finish()
+    showConfirmAlert(
+      this,
+      R.string.questionnaire_alert_back_pressed_message,
+      R.string.questionnaire_alert_back_pressed_title,
+      { finish() },
+      R.string.questionnaire_alert_back_pressed_button_title
+    )
   }
 }