+++ CHANGELOG.md
@@ -8,8 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - 2021-xx-xx
 
 ### Added
-- ANC | Added progerss bar to cql and measure reporting in ANC patien details page
-- ANC |Added measure reporting to ANC application
+- ANC | Optimized memory use for measure reporting by using String builder File forgotten in PR 646
+- ANC | Optimized memory use for measure reporting by using String builder and loading only once cql and measure reporting libraries
+- ANC | Added progress bar to cql and measure reporting in ANC patien details page
+- ANC | Added measure reporting to ANC application
 - Engine |Added class for Measure report evaluation which will be used in ANC application
 - ANC | Added Condition resource to sync params list
 - Moved Token to secure storage from AccountManager
@@ -22,6 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Quest | Feedback on UI
 - Quest | Remove custom activity
 - Quest | Add tags to patient via questionnaire
+- Engine | AlertDialog on questionnaire activity load, back, submit, progress
+- Engine | Block questionnaire submit if validation errors exist
 
 ### Fixed